Removed ZoodevioNode, using normal TreeNode
In trying to keep what we have to worry about down as well as getting
the folder ID onto the node so that you can find the folder id when
selecting nodes.  Name is on the TreeNode.Name, and Id is on
TreeNode.Tag.  The tag is just an object so the int ID must be converted
out using Convert.ToInt32(node.Tag).